pull any docker image

+++ scripts/start_worker.sh
@@ -1,6 +1,7 @@
 # !/bin/bash
 
 VERSION=0.1.7
+DEFAULT_IMAGE="ghcr.io/aicell-lab/bioengine-worker:$VERSION"
 WORKING_DIR=$(pwd)
 
 # Save all arguments
@@ -12,7 +13,6 @@ then
     echo "Apptainer could not be found. Please install it first."
     exit 1
 fi
-
 
 # Function to get argument value
 get_arg_value() {
@@ -61,32 +61,41 @@ set_arg_value() {
     fi
 }
 
-
 # Get the path to the image
-APPTAINER_IMAGE_PATH="$(get_arg_value "--image_path" "$WORKING_DIR/apptainer_images/bioengine-worker_$VERSION.sif")"
-APPTAINER_IMAGE_PATH=$(realpath $APPTAINER_IMAGE_PATH)
-set_arg_value "--image_path" $APPTAINER_IMAGE_PATH
+IMAGE="$(get_arg_value "--image" $DEFAULT_IMAGE)"
 
-# Set the directory and image name
-APPTAINER_IMAGE_DIR=$(dirname "$APPTAINER_IMAGE_PATH")
-APPTAINER_IMAGE=$(basename "$APPTAINER_IMAGE_PATH")
+# Get the image name and version
+if [[ "$IMAGE" == *.sif ]]; then
+    # Local Singularity image
+    IMAGE_PATH=$(realpath $IMAGE)
+    DOCKER_IMAGE=
+else
+    # Remote Docker image
+    FILE=${IMAGE##*/}
+    NAME=${FILE%%:*}
+    VERSION=${FILE##*:}
+
+    IMAGE_PATH="$WORKING_DIR/apptainer_images/${NAME}_${VERSION}.sif"
+    DOCKER_IMAGE=$IMAGE
+fi
 
 # Check if the BioEngine worker image is available
-if [ ! -f "$APPTAINER_IMAGE_PATH" ]; then
-    
-    VERSION=${APPTAINER_IMAGE#bioengine-worker_}
-    VERSION=${VERSION%.sif}
-    DOCKER_IMAGE=bioengine-worker:$VERSION
+if [ ! -f "$IMAGE_PATH" ]; then
+    if [[ -z "$DOCKER_IMAGE" ]]; then
+        echo "Error: Image file $IMAGE_PATH not found."
+        exit 1
+    fi
     # Ask user before pulling
-    read -p "Image file $APPTAINER_IMAGE_PATH not found. Do you want to pull it from Docker? (yes/no): " response
+    read -p "Image file $IMAGE_PATH not found. Do you want to pull it from Docker? (yes/no): " response
     if [[ "$response" =~ ^[Yy][Ee][Ss]$ ]]; then
-        echo "Pulling image $DOCKER_IMAGE from Docker..."
-        mkdir -p $APPTAINER_IMAGE_DIR
-        apptainer pull $APPTAINER_IMAGE_PATH docker://ghcr.io/aicell-lab/$DOCKER_IMAGE
+        echo "Pulling image $DOCKER_IMAGE..."
+        IMAGE_DIR=$(dirname "$IMAGE_PATH")
+        mkdir -p $IMAGE_DIR
+        apptainer pull $IMAGE_PATH docker://$DOCKER_IMAGE
         if [[ $? -eq 0 ]]; then
-            echo "Image pulled successfully to $APPTAINER_IMAGE_PATH"
+            echo "Successfully pulled $DOCKER_IMAGE to $IMAGE_PATH"
         else
-            echo "Error pulling image"
+            echo "Error pulling image $DOCKER_IMAGE"
             exit 1
         fi
     else
@@ -95,6 +104,7 @@ if [ ! -f "$APPTAINER_IMAGE_PATH" ]; then
     fi
 fi
 
+set_arg_value "--image" $IMAGE_PATH
 
 # Define environment variables
 ENV_VARS=()
@@ -133,45 +143,50 @@ add_bind() {
     fi
 }
 
-# Add SLURM-specific bindings
+# Add SLURM-specific bindings if SLURM is used
+MODE=$(get_arg_value "--mode" "slurm")
+if [[ "$MODE" == "slurm" ]]; then
+    # Binaries
+    add_bind $(which sinfo)
+    add_bind $(which sbatch)
+    add_bind $(which squeue)
+    add_bind $(which scancel)
 
-# Binaries
-add_bind $(which sinfo)
-add_bind $(which sbatch)
-add_bind $(which squeue)
-add_bind $(which scancel)
+    # Configuration files
+    add_bind "/etc/hosts"
+    add_bind "/etc/localtime"
+    add_bind "/etc/passwd"
+    add_bind "/etc/group"
+    add_bind "/etc/slurm"
+    add_bind "/etc/munge"
 
-# Configuration files
-add_bind "/etc/hosts"
-add_bind "/etc/localtime"
-add_bind "/etc/passwd"
-add_bind "/etc/group"
-add_bind "/etc/slurm"
-add_bind "/etc/munge"
+    # SLURM and Munge libraries
+    add_bind "/usr/lib64/slurm"
+    for lib in /usr/lib64/libmunge.so*; do
+        add_bind "$lib"
+    done
 
-# SLURM and Munge libraries
-add_bind "/usr/lib64/slurm"
-for lib in /usr/lib64/libmunge.so*; do
-    add_bind "$lib"
-done
-
-# Munge sockets
-add_bind "/var/run/munge"
-# Munge key
-add_bind "/var/lib/munge"
-# Munge logs
-add_bind "/var/log/munge"
+    # Munge sockets
+    add_bind "/var/run/munge"
+    # Munge key
+    add_bind "/var/lib/munge"
+    # Munge logs
+    add_bind "/var/log/munge"
+fi
 
 
 # Add BioEngine worker bindings
 
-# RAY_SESSION_DIR is needed by the Ray head node -> container path
-RAY_SESSION_DIR=$(get_arg_value "--ray_temp_dir" "/tmp/ray/$USER")
-RAY_SESSION_DIR=$(realpath $RAY_SESSION_DIR)
-mkdir -p "$RAY_SESSION_DIR"
-chmod 777 $RAY_SESSION_DIR
-add_bind $RAY_SESSION_DIR "/tmp/ray"
-set_arg_value "--ray_temp_dir" "/tmp/ray"  
+# If mode either SLURM or single-node:
+if [[ "$MODE" == "slurm" || "$MODE" == "single-node" ]]; then
+    # RAY_SESSION_DIR is needed by the Ray head node -> container path
+    RAY_SESSION_DIR=$(get_arg_value "--ray_temp_dir" "/tmp/ray/$USER")
+    RAY_SESSION_DIR=$(realpath $RAY_SESSION_DIR)
+    mkdir -p "$RAY_SESSION_DIR"
+    chmod 777 $RAY_SESSION_DIR
+    add_bind $RAY_SESSION_DIR "/tmp/ray"
+    set_arg_value "--ray_temp_dir" "/tmp/ray"
+fi
 
 # SLURM_LOGS_DIR is needed on the SLURM worker node -> real path
 SLURM_LOGS_DIR=$(get_arg_value "--slurm_logs" "$WORKING_DIR/slurm_logs")
@@ -224,16 +239,20 @@ fi
 # Cleanup function
 cleanup() {
     echo "Making sure the Ray head node is stopped..."
-    apptainer exec "$APPTAINER_IMAGE_PATH" ray stop --force
-    echo "Cleaning up any remaining Ray worker jobs..."
-    WORKER_JOB_IDS=$(squeue -u $USER -n "ray_worker" -h -o "%i")
-    if [[ -n "$WORKER_JOB_IDS" ]]; then
-        while read -r jobid; do
-            scancel $jobid
-        done <<< "$WORKER_JOB_IDS"
-        echo "All Ray worker jobs have been successfully canceled."
-    else
-        echo "No Ray worker jobs found to cancel."
+    apptainer exec "$IMAGE_PATH" ray stop --force
+
+    # If running in SLURM mode, cancel any remaining SLURM jobs
+    if [[ "$MODE" == "slurm" ]]; then
+        echo "Cleaning up any remaining Ray worker jobs..."
+        WORKER_JOB_IDS=$(squeue -u $USER -n "ray_worker" -h -o "%i")
+        if [[ -n "$WORKER_JOB_IDS" ]]; then
+            while read -r jobid; do
+                scancel $jobid
+            done <<< "$WORKER_JOB_IDS"
+            echo "All Ray worker jobs have been successfully canceled."
+        else
+            echo "No Ray worker jobs found to cancel."
+        fi
     fi
 }
 
@@ -246,5 +265,5 @@ apptainer exec \
     --pwd /app \
     "${ENV_VARS[@]}" \
     "${BIND_OPTS[@]}" \
-    "$APPTAINER_IMAGE_PATH" \
+    "$IMAGE_PATH" \
     python -m bioengine_worker "${BIOENGINE_WORKER_ARGS[@]}"
